This is the webscraper for amazon
You have to install puppeteer for this to work. Works for most products on amazon to web scrape a product you just have to paste the URL of the product into line 34 then run it by typing $node scrapers.js in the terminal. Then it will get the name of the product the price and the image. Working on why it doesn't work for certain products. Then will incorporate it into our google extension.